+++ src/lang/java.ts
@@ -116,12 +116,15 @@ export class JavaLangUtil extends LangUtil {
                         prepareImports();
                     }
                     let prevImport: [string, number] = ["", importStart];
-                    for (let i = 0; i < imports.length; i++) {
-                        const importContent = imports[i];
+                    for (const importContent of imports) {
                         const compareResult = importContent[0].localeCompare(rescue.value);
                         if (compareResult === 0) {
                             return;
                         }
+                    }
+                    for (let i = 0; i < imports.length; i++) {
+                        const importContent = imports[i];
+                        const compareResult = importContent[0].localeCompare(rescue.value);
                         if (compareResult > 0) {
                             // stop here
                             imports.splice(i, 0, [rescue.value, prevImport[1] + 1]);